Extract scatter plot code from Study Overview module
